requirements.txt
ansible
celery
django
django-celery
django-crispy-forms
redis
uwsgi
# optional
supervisor
